RSC convert admin doctors page
